my-first-app/CliffINDUS/cliffindus_backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
gunicorn
django-cors-headers
django-extensions